Move NamingConventions tests to new location
