+++ moodle/Tests/Sniffs/NamingConventions/fixtures/ValidFunctionName/validfunctionname_global.php
