+++ moodle/Tests/Sniffs/NamingConventions/fixtures/ValidFunctionName/validfunctionname_scoped.php
